Model building procedure implemented
We have now a well-commented procedure that is run, when the file is
executed.
Database class is still not implemented - database connection
still missing.
Next step is to check, if this would in general work (outsourced) and
afterwards adjust the code a little bit (est. 2 h) and write tests (est.
4 h). Apart from that classify.py is still missing, but that can be done
in 1 or 2 h, if build model runs.